# developer-docs/api-reference/integrations-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wareboxes Integration API",
    "description": "Version 1 partner integration surface for warehouse fulfillment workflows. Public operations are explicitly allowlisted; Wareboxes operator, RF, and administration endpoints are not part of this contract.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local development. Hosted environment URLs are supplied during integration onboarding."
    }
  ],
  "paths": {
...
    "/api/v1/portal/documents/{document_id}/content": {
      "get": {
        "tags": [
          "Visibility"
        ],
        "summary": "Download one immutable shipment document visible in the identity's scopes.",
        "operationId": "downloadCustomerPortalDocument",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "description": "Shipment document identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "x-wareboxes-tenant-id",
            "in": "header",
            "description": "Positive tenant context for the bearer credential.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable shipment document bytes.",
            "content": {
              "text/html": {}
            }
          },
          "401": {
            "description": "Missing or invalid credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The identity lacks customer portal permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The document is absent or outside the identity's owner/facility scopes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/portal/reports/inventory.csv": {
      "get": {
        "tags": [
          "Visibility"
        ],
        "summary": "Export the currently visible inventory availability as CSV.",
        "operationId": "exportCustomerPortalInventory",
        "parameters": [
          {
            "name": "inventory_owner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "search",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "include_history",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "x-wareboxes-tenant-id",
            "in": "header",
            "description": "Positive tenant context for the bearer credential.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scope-safe inventory availability CSV.",
            "content": {
              "text/csv": {}
            }
          },
          "401": {
            "description": "Missing or invalid credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The identity lacks customer portal permission or requested scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/portal/workspace": {
      "get": {
        "tags": [
          "Visibility"
        ],
        "summary": "Read the customer-facing inventory and fulfillment workspace.",
        "operationId": "getCustomerPortalWorkspace",
        "parameters": [
          {
            "name": "inventory_owner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "search",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "include_history",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "x-wareboxes-tenant-id",
            "in": "header",
            "description": "Positive tenant context for the bearer credential.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner- and facility-scoped customer visibility projection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPortalWorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The identity lacks customer portal permission or requested scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CustomerPortalDocumentResponse": {
        "type": "object",
        "required": [
          "document_id",
          "shipment_id",
          "order_id",
          "order_key",
          "inventory_owner_id",
          "facility_id",
          "document_type",
          "file_name",
          "media_type",
          "content_length",
          "content_sha256",
          "generated_at",
          "download_path"
        ],
        "properties": {
          "content_length": {
            "type": "integer",
            "format": "int64"
          },
          "content_sha256": {
            "type": "string"
          },
          "document_id": {
            "type": "integer",
            "format": "int64"
          },
          "document_type": {
            "$ref": "#/components/schemas/CustomerPortalDocumentType"
          },
          "download_path": {
            "type": "string"
          },
          "facility_id": {
            "type": "integer",
            "format": "int64"
          },
          "file_name": {
            "type": "string"
          },
          "generated_at": {
            "type": "string"
          },
          "inventory_owner_id": {
            "type": "integer",
            "format": "int64"
          },
          "media_type": {
            "type": "string"
          },
          "order_id": {
            "type": "integer",
            "format": "int64"
          },
          "order_key": {
            "type": "string"
          },
          "shipment_id": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "CustomerPortalDocumentType": {
        "type": "string",
        "enum": [
          "packing_slip",
          "carton_label_set"
        ]
      },
      "CustomerPortalInventoryResponse": {
        "type": "object",
        "required": [
          "inventory_owner_id",
          "inventory_owner_name",
          "facility_id",
          "facility_name",
          "item_id",
          "uom",
          "status",
          "on_hand",
          "reserved",
          "held",
          "available"
        ],
        "properties": {
          "available": {
            "type": "integer",
            "format": "int64"
          },
          "expiration": {
            "type": [
              "string",
              "null"
            ]
          },
          "facility_id": {
            "type": "integer",
            "format": "int64"
          },
          "facility_name": {
            "type": "string"
          },
          "held": {
            "type": "integer",
            "format": "int64"
          },
          "inventory_owner_id": {
            "type": "integer",
            "format": "int64"
          },
          "inventory_owner_name": {
            "type": "string"
          },
          "item_description": {
            "type": [
              "string",
              "null"
            ]
          },
          "item_id": {
            "type": "integer",
            "format": "int64"
          },
          "lot": {
            "type": [
              "string",
              "null"
            ]
          },
          "on_hand": {
            "type": "integer",
            "format": "int64"
          },
          "primary_sku": {
            "type": [
              "string",
              "null"
            ]
          },
          "reserved": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "uom": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CustomerPortalOrderResponse": {
        "type": "object",
        "required": [
          "order_id",
          "order_key",
          "inventory_owner_id",
          "inventory_owner_name",
          "status",
          "rush",
          "ordered_quantity",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "destination_city": {
            "type": [
              "string",
              "null"
            ]
          },
          "destination_company": {
            "type": [
              "string",
              "null"
            ]
          },
          "destination_country": {
            "type": [
              "string",
              "null"
            ]
          },
          "destination_region": {
            "type": [
              "string",
              "null"
            ]
          },
          "facility_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "facility_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "inventory_owner_id": {
            "type": "integer",
            "format": "int64"
          },
          "inventory_owner_name": {
            "type": "string"
          },
          "order_id": {
            "type": "integer",
            "format": "int64"
          },
          "order_key": {
            "type": "string"
          },
          "ordered_quantity": {
            "type": "integer",
            "format": "int64"
          },
          "rush": {
            "type": "boolean"
          },
          "ship_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/CustomerPortalOrderStatus"
          }
        },
        "additionalProperties": false
      },
      "CustomerPortalOrderStatus": {
        "type": "string",
        "enum": [
          "open",
          "held",
          "processing",
          "awaiting_packing",
          "packing",
          "awaiting_shipment",
          "shipped",
          "cancelled",
          "void"
        ]
      },
      "CustomerPortalShipmentResponse": {
        "type": "object",
        "required": [
          "shipment_id",
          "order_id",
          "order_key",
          "inventory_owner_id",
          "inventory_owner_name",
          "facility_id",
          "facility_name",
          "status",
          "carton_count",
          "shipped_quantity",
          "tracking_numbers",
          "created_at"
        ],
        "properties": {
          "carrier": {
            "type": [
              "string",
              "null"
            ]
          },
          "carton_count": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string"
          },
          "departed_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "facility_id": {
            "type": "integer",
            "format": "int64"
          },
          "facility_name": {
            "type": "string"
          },
          "inventory_owner_id": {
            "type": "integer",
            "format": "int64"
          },
          "inventory_owner_name": {
            "type": "string"
          },
          "manifested_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "order_id": {
            "type": "integer",
            "format": "int64"
          },
          "order_key": {
            "type": "string"
          },
          "service": {
            "type": [
              "string",
              "null"
            ]
          },
          "shipment_id": {
            "type": "integer",
            "format": "int64"
          },
          "shipped_quantity": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "$ref": "#/components/schemas/CustomerPortalShipmentStatus"
          },
          "tracking_numbers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "CustomerPortalShipmentStatus": {
        "type": "string",
        "enum": [
          "awaiting_manifest",
          "manifested",
          "partially_departed",
          "departed",
          "cancelled"
        ]
      },
      "CustomerPortalWorkspaceResponse": {
        "type": "object",
        "required": [
          "generated_at",
          "inventory",
          "orders",
          "shipments",
          "documents",
          "inventory_report_path"
        ],
        "properties": {
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerPortalDocumentResponse"
            }
          },
          "generated_at": {
            "type": "string"
          },
          "inventory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerPortalInventoryResponse"
            }
          },
          "inventory_report_path": {
            "type": "string"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerPortalOrderResponse"
            }
          },
          "shipments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerPortalShipmentResponse"
            }
          }
        },
        "additionalProperties": false
      },
      "ErrorReason": {
        "type": "string",
        "description": "Stable machine-readable reason for an API failure.",
        "enum": [
          "unauthorized",
          "forbidden",
          "not_found",
          "validation_failed",
          "invalid_request",
          "conflict",
          "invalid_state_transition",
          "insufficient_inventory",
          "idempotency_key_required",
          "idempotency_key_reused",
          "precondition_required",
          "revision_conflict",
          "invalid_cursor",
          "method_not_allowed",
          "payload_too_large",
          "unsupported_media_type",
          "rate_limited",
          "internal_error"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error response returned by every version 1 endpoint.",
        "required": [
          "reason",
          "message",
          "request_id"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "idempotency key was already used with a different request"
          },
          "reason": {
            "$ref": "#/components/schemas/ErrorReason"
          },
          "request_id": {
            "type": "string",
            "example": "partner-order-1001-attempt-1",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]{1,128}$"
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            }
          }
        },
        "additionalProperties": false
      },
      "FieldViolation": {
        "type": "object",
        "description": "Machine-addressable validation failure for one request field.",
        "required": [
          "field",
          "reason"
        ],
        "properties": {
          "field": {
            "type": "string",
            "description": "Request field or logical input that failed validation.",
            "example": "lines[0].quantity"
          },
          "reason": {
            "type": "string",
            "description": "Stable validation explanation intended for an integration operator.",
            "example": "must be a positive integer"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque bearer credential provisioned for the integration identity. It is not a JWT and must be stored as a secret."
      }
    }
  },
...
}